Fully Working Locally Containerized and Uncontainerized without games.json or the GameService.cs
