+++ Bucstop WebApp/BucStop/Controllers/GamesController.cs
@@ -16,22 +16,20 @@ namespace BucStop.Controllers
     {
         private readonly MicroClient _httpClient;
         private readonly PlayCountManager _playCountManager;
-        private readonly GameService _gameService;
         private readonly ILogger<GamesController> _logger;
-        private Task<List<Game>> gamesAsync; 
+        private Task<List<Game>> gamesAsync;
 
 
-        public GamesController(MicroClient games, IWebHostEnvironment webHostEnvironment, GameService gameService, ILogger<GamesController> logger)
+        public GamesController(MicroClient microClient, IWebHostEnvironment webHostEnvironment, ILogger<GamesController> logger)
         {
-            _httpClient = games;
-            _gameService = gameService;
+            _httpClient = microClient;
             _logger = logger;
 
-            // Initialize the PlayCountManager with the web root path and the JSON file name
-            _playCountManager = new PlayCountManager(_gameService.GetGames() ?? new List<Game>(), webHostEnvironment);
-
             //start the async pull of the games info
-            gamesAsync = GetGamesWithInfo();
+            gamesAsync = _httpClient.GetGamesWithInfo();
+
+            // Initialize the PlayCountManager with the web root path and the JSON file name
+            _playCountManager = new PlayCountManager(_httpClient.GetGamesList() ?? new List<Game>(), webHostEnvironment);
         }
 
         //Takes the user to the index page, passing the games list as an argument
@@ -102,58 +100,6 @@ namespace BucStop.Controllers
 
             return View(game);
         }
-
-        public async Task<List<Game>> GetGamesWithInfo()
-        {
-
-            List<Game> games = new List<Game>();
-
-            try
-            {
-                GameInfo[] gameInfos = await _httpClient.GetGamesAsync();
-
-                if (gameInfos.Length > 0)
-                {
-                    _logger.LogInformation("Successfully retrieved {Count} games from API.", gameInfos.Length);
-                }
-                else
-                {
-                    _logger.LogWarning("API returned 0 games.");
-                }
-
-                foreach (GameInfo info in gameInfos)
-                {
-                    Game game = new Game();
-
-                    if (info != null)
-                    {
-                        game.Id = info.Id;
-                        game.Title = info.Title;
-                        game.Content = info.Content;
-                        game.Thumbnail = info.Thumbnail;
-                        game.Author = info.Author;
-                        game.HowTo = info.HowTo;
-                        game.DateAdded = info.DateAdded;
-                        game.Description = $"{info.Description} \n {info.DateAdded}";
-                        game.LeaderBoard = info.LeaderBoard;
-
-                        _logger.LogInformation("Game ID {Id} Content URL: {Content}", info.Id, info.Content);
-
-                    }
-
-                    games.Add(game);
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Error retrieving game information from API.");
-            }
-
-
-            return games;
-        }
-
-
 
         //Takes the user to the deprecated snake page
         public IActionResult Snake()
