--- Bucstop WebApp/BucStop/games.json
@@ -1,26 +1,0 @@
-[
-  {
-    "Id": 1,
-    "Title": "Snake",
-    "Content": "/js/snake.js",
-    "Description": "Snake Description",
-    "Thumbnail": "/images/snake.jpg",
-    "PlayCount": 0
-  },
-  {
-    "Id": 2,
-    "Title": "Tetris",
-    "Content": "/js/tetris.js",
-    "Description": "Tetris Description",
-    "Thumbnail": "/images/tetris.jpg",
-    "PlayCount": 0
-  },
-  {
-    "Id": 3,
-    "Title": "Pong",
-    "Content": "/js/pong.js",
-    "Description": "Pong Description",
-    "Thumbnail": "/images/pong.jpg",
-    "PlayCount": 0
-  }
-]
